+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Header from "@/app/components/Header";
+import {Header} from "@/block";
 
 const Home = () => {
