+++ src/block/index.ts
@@ -1,0 +1,3 @@
+import Header from "@/block/header/Header";
+
+export {Header};
